- Split up the s7 driver code into multiple files to be better manageable.
